Update Typing Tutor.py
This is a Typing Tutor program. In this program, you are given a string which you have to type. After submitting it (to submit you have to press 's') it will show you the seconds you took to type, your typing speed (in WPM i.e. Words Per Minute), number of correct and wrong words and your accuracy. This program was created on 18th September, 2020. Due to some issues, it is being posted so late i.e. 13th October, 2020. This program is still in progress. Hope you like it.